requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
boto3
botocore
openai
transformers
tk
python-dotenv